- osc.httprequest: added build_url module method
build_url provides a way to construct an url from an apiurl (http://host)
a path and a query str. This method should be used if a url has to be
constructed manually (this is only needed in very rare cases)

Urllib2HTTPRequest uses this method too (refactored some code).