- osc.cli: added cli package
Added description module which can be used to define a argparse
ui in a descriptive manner. The syntax for defining a command is
similar to defining models in the django framework (that's what
inspired me).

Example:

OscCommand is a base class for all toplevel osc commands.

class Request(CommandDescription, OscCommand):
    """Show and modify requests."""
    cmd = 'request'

class RequestList(CommandDescription, Request):
    """List requests.

    By default open requests for a specific project or package will be
    listed.

    Examples:
    osc request list api://
    osc request list api://project
    osc request list api://project/package

    """
    cmd = 'list'
    args = 'api://project?/package?'
    opt_user = Option('U', 'user', 'list only requests for USER')
    opt_group = Option('G', 'group', 'list only requests for GROUP')
    opt_state = Option('s', 'state', 'list only requests with state STATE',
                       choices=['new', 'review', 'accepted', 'revoked',
                                'declined', 'superseded'], action='append',
                       default=['new', 'review'])
    func = call(request.list)

That is with the help of the inheritance hierarchy we can build a
command <-> subcommand hierarchy. For more details and examples
have a look at the following commits or at this blog post
http://lizards.opensuse.org/2012/08/07/gsoc-osc2-client-summary-of-week-11/