- test.oscargs: added testcases to handle name clashes
A name clash occurs if multiple oargs are specified with the same
"name": "repo/arch repo/xyz" leads to a name clash. By default
the argument bound to "repo/xyz" is stored in the info's repo
attribute. If "ignore_clashes=False" is passed to the OscArgs' __init__
method both attributes (the one bound to "repo/arch" and "repo/xyz")
will be stored in a list to which the info's repo attribute refers to.