- osc.cli.util.shell: added class ItemSelector
An item selector can be used to select one out of multiple
items. It will prompt for an input and the item which corresponds
to the input will be returned.
This class will be used by the "review shell" in order to select a
specific review which should be accepted, declined etc.

Later this class will also be used to select a accept, decline etc.
message from a list of (user specified) accept, decline etc. messages
(for this task its <tab>-completion feature might come handy...)