- osc.cli.util.shell: added class AbstractItemSelectorFactory and restructured classes which are used by the ItemSelector
The main purpose of the factory is to create an ItemSelector
(or subclass) instance. Additionally it can be used to hide
creation of the objects which are used by the item selector (like
the "item storage", a "wrapper" for the items etc.).
This simplifies the code in the ReviewShellController (not yet committed)
a lot.