- util.io: added TemporaryDirectory class
The TemporaryDirectory class is inspired by lnussel's mytmpdir
class in osc's (osc1) build.py module.

An instance of this class represents a temporary directory in
the filesystem.
The temporary directory is created lazily. Once an instance of this
class is destroyed, the corresponding temporary directory is removed
as well (unless delete=False was passed to the __init__ method).
Moreover, an instance of this class can be used as a
context manager (__exit__ deletes the temporary directory, unless
delete=False was passed to the __init__ method).