- util.delegation: added initial delegation module
It provides classes that facilitate delegation tasks.

These classes can be used, if normal duck typing is not sufficient
anymore. Suppose an instance of the class osc2.util.io.TemporaryDirectory
should be passed to the os.stat function. This will not work, because os.stat
requires a str/unicode or buffer instance as a parameter. Thus, simple
duck typing is insufficient (PyUnicode_FromEncodedObject, which is
indirectly called by posix_do_stat, has "strict" type checks).
So the idea is to wrap the TemporaryDirectory instance in an instance that
behaves like/is a str/unicode instance. Additionally, all non-str method
calls are delegated to the real TemporaryDirectory instance.
For this, this module provides several classes.

The use-case from above is implemented in the osc2.util.io.mkdtemp function.