- test.util.test_io: test corner case in TemporaryDirectory.__del__
Avoid an unraised exception:
Exception AttributeError: "'NoneType' object has no attribute 'path'"
in <bound method TemporaryDirectory.__del__ of
<osc2.util.io.TemporaryDirectory object at 0x7f8abc8d7910>> ignored

This exception was "raised" in the TemporaryDirectory.__del__ method,
if the "osc2.util.io" was already "unloaded" (the __del__ method
calls the _cleanup method, which tried to call the
os.path.isdir function).
What happened is the following:
If the "osc2.util.io" module is passed to _PyModule_Clear
(Objects/moduleobject.c), almost all entries in the module dict are
set to None. That is, the entry for the key "os" points to None.
Now, suppose that there is still a TemporaryDirectory instance
around, whose __del__ method is eventually called. During this call,
"os" resolves to None and therefore the subsequent lookup for the
name "path" leads to the AttributeError.
Instead of raising this exception, it is just printed to stderr,
because it occurred in a __del__ method (see slot_tp_del
(Objects/typeobject.c) and PyErr_WriteUnraisable (Python/errors.c)).