Document an issue with the wc locking code
Unlinking the lock file yields to races and we can end up in a
situation where two processes think that they have exclusive access
to the package wc.
The fix is easy (do not unlink the lock file), but we probably have
to adjust the testcases as well (I have no time for it atm).